Migrate Button to Styled Components (#333)
