+++ src/button/specifications/secondary.md
@@ -17,7 +17,7 @@ After tapping if request is already loaded we either go to the next page or clos
 If request is not loaded we display the loading state of these buttons. When loaded, we either go to the next page or close the modal or if it’s a final decision we display the checked state.
 
 ```js
-<Button status={Button.STATUS.SECONDARY}>content</Button>
+<Button status={ButtonStatus.SECONDARY}>content</Button>
 ```
 
 ## When should I use it?
